Remove check_sudo_usage logic from tox
This doesn't make much sense in an ansible only world. There is my
ways today to block sudo or check if sudo failed with ansible. No need
to include our custom bash script for this check.

If a job wants sudo, they can either not use the revoke-sudo role or
write another job.

Change-Id: I267541a813b5ff6e7301f50421af46a0db33304d
Signed-off-by: Paul Belanger <pabelanger@redhat.com>