Remove OOM check from tox role
This is something we did in openstack-infra but not sure it make sense
for default tox jobs.  It is possible we want to make this an ansible
role for playbooks to opt into post-run. But with properly logging on
a job, it should be straightforward to determine why a tox test
failed.

Change-Id: I746fae586a3df40a07ceca9c0bebe579f9f74f34
Signed-off-by: Paul Belanger <pabelanger@redhat.com>