Create tox_environment_defaults variable for tox based jobs
Expose the ability for tox jobs to define a default set of
environmental variables to pass to the tox role.  We also do a few new
things.

Check to see if tox_environment and tox_environment_defaults are
defined, if so combined tox_environment_defaults with
tox_environment creating a new my_tox_environment variable.
Variables defined in tox_environment take priority when we combine
variables.

Change-Id: Ic4a5644e79ca7000cbdaee40467619a6832c8332
Signed-off-by: Paul Belanger <pabelanger@redhat.com>