Add fetch-stestr-output role
This commit adds a new role for jobs that run stestr instead of
testrepository. This will generate the output using stestr for those
cases like we do with testr for the old jobs which leverage
testrepository. This change is the dual of the jenkins script patch:
I1fc3b146e99e3484fa009b227ee5d0aeb6ed6084

Change-Id: Ibf652801507ad2845281dd540b4007b04875b2ae