Remove base job
With I26ffddad8358c156cfac749ce98af70f3447f671, base jobs may no
longer be defined in untrusted projects, and we want zuul-jobs
to be usable as an untrusted project.  Remove the base job from
here; later we can add it to zuul-base-jobs and treat that as a
config repo.

Change-Id: I9688fd844776c8fc5cf9bdd555ec3bcde7ead054
Depends-On: I26ffddad8358c156cfac749ce98af70f3447f671