Add a role to remove an ssh private key
A role was added to add a private key. This role is a companion that
will remove the private key, as an extra precaution around the secret.
There is a non-zero chance the nodepool resource will not get properly
cleaned up, so we want to ensure that we do our best to clean out the
secrets.

Change-Id: Ib365b2d9304d7ccdc03df97b1d2ad924d6e8513b