Handle floating ips in multi-node-known-hosts
Nodes (and ansible) don't know about floating ips because they are
external. However, they are valid interfaces over which nodes might talk
to each other- especially if we're re-using the inventory generated for
the jobs.

Start off by grabbing all of the addresses from nodepool, if they exist.
Then do the loop for finding the addresses we want. Then collapse
that list with set and do a pass to create the actual list of hosts and
keys needed.

Change-Id: I748fae1c86d7bb9333dc251991308282858811a4