Disable deltarpm for CentOS and Fedora
deltarpm is useful when the bottleneck is the network throughput
on long-lived instances where you need to *update* packages that
are already installed.
It also requires the mirrors we are using to mirror the drpm packages
which is not currently done in our infrastructure.

Let's disable deltarpm as it's not suited to our use case.

Change-Id: Ibf3b6093496375efda26f46eaf90438af0adb394