Handle projects without setup.py or valid setup.cfg
project-config has tox jobs but does not have a setup.py. This confuses
the siblings code.

First - check for a setup.cfg and if one is not there just give up.

Second - we switched to just reading setup.cfg in the other part of the
module. Do that for determining package name as well.

Change-Id: I1f33261871791863e77789b522fc6f7db3f3cdbb