Always create dest dir in zuul-cloner shim
The "cp" command copies the source files into the destination
directory.  The zuul-cloner shim relied on this behavior and so
specified one level above the intended destination as the destination
directory, so that cp would end up creating the final directory based
on the final component of the source directory.

This did not behave as expected if the destination directory was not
intended to be the name of the project.  For example, some uses clone
directly into the workspace directory.  Others may be intentionally
altering the name of the destination directory (eg, to strip "puppet-"
from the project name).

To correct this, always ensure the full destination directory exists,
and always copy the entire source directory contents *into* the
destination directory.

Change-Id: If80df82b73c59a3c81e0c8ed5ff0d135dd5c26c9