zuul-cloner: enter directory before copying
The "-a" argument to cp implies "--preserve=all" which changes the
permissions of the target directory; e.g.

 mkdir /tmp/foo
 chmod 700 /tmp/foo
 cp -al /some/dir/. /tmp/foo
 ls -l /tmp/foo

and you will find that /tmp/foo is probably not 700 any more.  This
causes problems when /tmp/foo is a home directory, because it can make
them world-writable and hence ssh won't log into them any more.

For simplicity, switch into the directory we want to copy into, and
then copy everything in.  This should stop cp touching the parent
directory in any way.

Change-Id: Icd842b6fba220d41601adeaa0c5a41b2dab582bc