Make fetch-tox-output more resilient
The current find command fails when there is an empty log directory
in the dist dir as seen here:

  http://logs.openstack.org/87/508287/1/check/openstack-tox-py27/4e00cb9/job-output.txt.gz#_2017-09-29_04_10_39_589261

Update it to make it look specifically for the tox environments
specified in tox_envlist.

Make sure it supports a comma separated list as well as 'all' in
tox_envlist, as those are stated valid inputs.

Depends-On: I1c98508fe227f7f6aa1d0b2f7dcf270cecaa60d8
Change-Id: Id28dbd19f4f612e3e9b9feb6350250b18ae2f8c6