Add a generic process-test-results role
Takes an stestr or testr repo as input and generates a subunit
file and an html report out of it. The files are stored in the
staging area on the test node.

This in combination with a generic role to pull the staging
folder to the zuul executor is meant to eventually replace to
existing roles:
- fetch-testr-output
- fetch-stestr-output

Change-Id: Id6149d4e265ab9f0ab6d8faeffdec651c63dc056