Add a toggle to disable ARA static report generation
Without this patch, it is not possible for jobs to disable the
ara report generation because it is part of the base job by default.

It might be useful (or required) for jobs to disable the static
report generation if there are issues with it or if the static report
is causing problems. This toggle will provide that ability.

Change-Id: Ia5e43db9f852d405ea987b380e36cb93c983314d