Revert "Revert "Add zuul.{pipeline,nodepool.provider,executor.hostname} to job header""
The nodepool provider variable is not always available, for example
when running only on localhost, so default it to "N/A".

Additionally, we don't want these tasks to run multiple times if we
are running a multi-node job so use "run_once: yes" to prevent
printing the header more than once.

This reverts commit 27755ba83f83858f3dec3c23c3ab5101ea93a692.

Depends-On: Ia79a793200fcb89161783ff641b85106936084b5
Change-Id: Id9117935d9c53268ea893796154af03db21ee3a7