Move to dictionary list of projects zuul._projects (take 2)
This is a revert of commit f28cc33bc3bc7f8a30d868964ec7712c26cef8f9
which was a revert of the original attempt at moving to zuul._projects
(df987735730e712ea3bc57f99d3c96cb561c43a9).

The issue was that the order of projects in "Determine local HEAD" was
not the same as "Update remote repository state" and we ended up
checking out the wrong thing.

To avoid this, build the projects and their current heads on the
exectutor into a separate dict, and as we iterate the projects look up
the correct values in that.

Change-Id: I0d9a66435714efe42fb653aafc6cb71b08ca6b8a