Ensure ChangeLog exists in install-if-python
We install projects if they are python in the sphinx jobs because
autodoc needs to be able to import the package. Also, for pbr projects,
many projects have links to ChangeLog in the docs, which is generated by
pbr on install and sdist. Installing with pip install --no-deps does not
cause the ChangeLog to be generated, so trigger it via an sdist build
first.

Change-Id: I4528cc7f322d1fc753eb5bc259f70c47ee9d1d80