Do not use --ignore-missing-args for rsync
rsync --ignore-missing-args is only supported with rsync 3.1. CentOS 7
uses rsync 3.0 and fails.

We ignore missing remote files, so remove the argument and let the role
ignore an error if the files do not exist.

Update fetch-tox-output to use failed_when as well. We really want the
role not to fail (failed_when) instead of showing a failure and then the
playbook ignore it (ignore_errors).

Change-Id: I24bfb5519acccd7fd36b4839bb598436b4c8f01c