Move revoke-sudo from unittest pre to job run
Having revoke-sudo in the pre playbook of unittest has made us not able
to re-use unittest for javascript, and also means we can't use
nodejs-npm for nodejs-npm-run-test, which is a bit odd.

Instead, add the role to the run playbook for tox and npm, which allows
us to stack and combine things a bit better.

Change-Id: I0c7c508996ad2647db706d4ebf156c5dd96bf9b5