Adjust check for .stestr directory
stestr can leave a .stestr directory from which subunit cannot be
generated after an initial run that fails due to a python import/parse
error.

In order to determine if a run will be able to generate subunit output,
check for the '.stestr/failing' file.

NOTE: This is **NOT** an indication of a failing run. It's a file that
will be present in the .stestr directory if an entire run has happened
and acts as a marker that the directory has enough information to export
a subunit stream.

Change-Id: Id1dd696c81da09e9d2e5fe350337b73eb0646d2d