Don't fetch coverage report from failing tests
When test runs fail, coverage reports are not generated. We try to fetch
them anyway, which leads to POST_FAILURE instead of FAILURE.

Add in a check to ensure the test run was successful, then a check for
the coverage dir and an explicit error message so that it's clear to
people that the issue is that coverage wasn't generated after a
successful run, rather than there being an issue with rsync or
something.

Change-Id: I07f942a9df518d6602ee151d82b35ce413769c73