Add no_log to verbouse tasks
Two tasks in stage-output "Register sources" and "Set source and
destination for files and folders" spam the whole output of "stat"
to be printed to console for each and every file to be staged.
The info is printed twice and it's hardly valuable for debugging purposes.

The following task, "Build a list of source, dest dictionaries for text
file", prints a more compact list of all files and folders that are going
to be staged, which I believe is enough for debugging.

To solve this add no_log to the two tasks mentioned above.

Change-Id: Id52047c429233c6684f2fca5537674f91fb2603d