Setup repositories for openSUSE Tumbleweed
Tumbleweed as a rolling release model distribution does not
have a sensible release version and is kept distinct from the
classing "release + update" distributions that are called "Leap".

Conditionalize the repository setup for tumbleweed properly,
the base path is different and it does not have a releaseversion
component (the release version is always changing, it is
the snapshot date in the form YYYYMMDD).

Change-Id: I700ba2279dc7aadc7051cad53898e1de9007a08d