Fix lookup of subunit files in fetch-subunit-output
The role allows override of the zuul_work_dir variable, however in the
step where it retrieves the test results is the only step where it
prepends the directory with a fixed path. This breaks the step if the
zuul_work_dir given is a fully qualified path.

This patch defaults the zuul_work_dir to a fully qualified directory and
standarizes the usage zuul_work_dir across the role.

Change-Id: I62bc7571f625586a0d1b660f896e0c7cbaeae6e1