Add known hosts for all users
The role today only adds known hosts for the zuul user.
This replicates what d-g does today by adding a default known_hosts
file which applies to all users.

We csnnot remove the original task that adds for the zuul user since
there's a unit test in openstack-zuul-jons that expects that.

Depends-on: https://review.openstack.org/550425/

Change-Id: I54d4d975fa54cbb6df7dbed525211ab5c2cd6754