Add a job variable to control javascript link copying
The zuul multi-tenant dashboard job wants to make a few symlinks to
simulate what would happen on a real deployment with rewrite rules. The
current hard-coded behavior of copying the links causes ... havoc.

Add a job variable defaulting to the current behavior which such jobs
can flip if they need to.

Change-Id: I751524baa80353606d74b87dbaacfe7cd8b71a5e