Log output and exception when traceroute fails
These traceroutes currently fail in a very opaque way.  We are
occasionally seeing Fedora nodes fail in here (which is odd, because
obviously networking is up enough for zuul to connect) and don't have
much to go on.

When an exception is caught, add the output, return code and basic
traceback to the return attributes, and include them in the failure
case.

Depends-On: https://review.openstack.org/563702
Change-Id: I047bf2b1daa22a5b6bfc12b3f42b108975097409