Add role to trigger readthedocs via authenticated endpoint
The old API endpoint is deprecated, so add a role that pings the
webhook as described in [1].

Authentication can happen via a token or username/password combo,
which need to be kept in secrets.

The webhook api has an "id" value which some fiddling suggests is an
incrementing integer each time a webhook endpoint is generated (if you
create, delete, create your webhook via the admin interface, you often
as not get the next value).  Unfortunately, it appears you can not
query this in any way automatically; it has to be provided.  However,
I do not believe this has to be secret; you can not trigger the
end-point without authentication (either password or token).

This means the generic playbook here will not work any more.  Because
secrets are bound to the playbook they are defined within, this role
will need to be called directly.  It will be removed in a follow-on.

[1] https://docs.readthedocs.io/en/latest/webhooks.html

Change-Id: I651efdb093df85cea3ab2eaf1a5a9256c87a2ca4