Install build bindep profiles alongside doc and test
In order to be able to separate out depends needed for building wheels,
depends needed for running tests and depends needed for running
software, we can tag build depends with a 'build' profile. However,
doing that then means we don't have all of the things we need for
building docs or testing things.

Add build to the profile list for both docs builds and test installs.

Change-Id: I0689acb53789d3650926a3e11aec4570df5d950c
Needed-By: https://review.openstack.org/580159