Add role for installing docker and configuring registry mirror
There are two different ways to install docker - from upstream repos or
from distro. At the moment each of these carries with it the need for a
different caching proxy setup.

Add a role that will install docker from upstream by default, but which
also supports installing directly from distros. The role also sets up
the registry proxy appropriately for each.

This role at the moment only works on ubuntu. It should obviously be
updated to work on centos and fedora as well.

Needed-By: https://review.openstack.org/580160
Change-Id: I7d6bac68a2c0fecf13a8bd9535a3fdeb85e7d999