Enable options for generating the parent links
This change restores the --no-parent-links option that was dropped at some
point.

It also adds a new --create-topdir-parent-link option which defaults to
False. This is a separate option to hte parent links as it creates a
relative link to '../' from the results of the logs. This makes sense if it
is going back to some kind of global index of all the job results, but in
most cases if logs are in containers behind some kind of CDN it's likely
the entrypoint into the logs is going to be from the zuul dashboard (or
comments in the review system). Thus default to False.

Change-Id: I3b5ef4ac81bd7dbc9e65a5d22d81198b6642fa58