Debugging for readthedoc web ping
Unfortunately this is failing in the gate, and I have no idea why (it
works locally ...).  Add some debugging.  Since this is called with a
secret from a post-review only job in the trusted project-config repo,
there's little way to self-test this without committing things :/
Hopefully this sheds some light ...

Change-Id: I7d661f8eefe25d90b5eded309062f32344b3379c