Remove mimetype check when calculating size
We currently try to skip size formatting for folders. However we
compare with a bogix mimetype so the check is false in every case.
Further folders typically have a size of 512 bytes or 4k. In both
cases we don't really need to skip the size formatting so instead of
fixing the check just skip it and do the size formatting
unconditionally.

Change-Id: I7ef021381bb56acf4b22551cc5d5613470fd6d08